refactor: 스크린 경로와 하단 네비게이션 아이템 분리하여 관리
