+++ app/src/main/java/com/example/mz_focusnews/core/components/BottomNavItem.kt
@@ -6,15 +6,6 @@ import com.example.mz_focusnews.core.theme.myiconpack.Home
 import com.example.mz_focusnews.core.theme.myiconpack.Quiz
 import com.example.mz_focusnews.core.theme.myiconpack.User
 
-sealed class ScreenRoute(val route: String) {
-    object Content : ScreenRoute("content/{newsId}")
-    object Like : ScreenRoute("like")
-    object Recent : ScreenRoute("recent")
-    object QuizIntro : ScreenRoute("quiz_intro")
-    object QuizPlay : ScreenRoute("quiz_play")
-    object Ranking : ScreenRoute("ranking")
-}
-
 sealed class BottomNavItem(
     val title: String,
     val icon: androidx.compose.ui.graphics.vector.ImageVector,
